many-to-many sites
Make the relationshio between water allocations amounts and sites as many-to-many to accomodate California's request.

Make Date fields in the fact tabels as IDs because they are foreign keys not actual dates.

Delete the Primary beneficial use CV table because now we're reusing the main beneficial use CV table for that purpose